Several deletions for useles archives. Checklist controller created and routes changed, items not working
