+++ app/controllers/api/v1/users_controller.rb
@@ -19,74 +19,6 @@ module Api
         render json: @user, status: :created
       end
 
-
-      def create_checklist
-        user = User.find(params[:user_id])
-        checklist = Checklist.create!(name: params[:name], user_id: params[:user_id])
-        categories_type = CategoryType.all
-        items_type = ItemType.all
-        categories_type.each do |category_type|
-          category = Category.create!(checklist_id: checklist.id, category_type_id: category_type.id)
-
-          item_array = items_type.where(category_type_id: category_type.id.to_s)
-          item_array.each do |item|
-            item = Item.create!(category_id: category.id, item_type_id: item.id.to_s, status: "not done")
-          end
-        end
-        render json: "Checklist Criada"
-      end
-
-
-
-      def checklists
-        id = params[:id]
-        checklist = Checklist.where(user_id: id)
-        render json: checklist
-      end
-
-      def checklist
-        id = params[:id]
-        checklist = Checklist.find(id)
-        categories = Category.where(checklist_id: id)
-        category_type = CategoryType.all
-        items_type = ItemType.all
-        items = Item.all
-        count = 0
-        categories_body = []
-        a = []
-        categories.each do |category|
-          category.items.each do |i|
-            item_count = 0
-            item = items_type.where(id: i.item_type_id)
-            i.name = item[item_count].name
-            item_count += 1
-          end
-
-          categories_body << category
-          name = category_type.where(id: category.category_type_id)
-          categories_body[count][:name] = name[0].name
-          count += 1
-        end
-        render json: categories_body
-      end
-
-      def delete_checklist
-        @id = params[:id]
-        checklist = Checklist.where(_id: @id)
-        checklist.destroy
-        render json: "Checklist deletada"
-      end
-
-      def item
-        id = params[:id]
-        @item = Item.find(id)
-        if params[:status] == "done" || params[:status] == "not done" || params[:status] == "partially done"
-          @item.update!(status: params[:status])
-        end
-        render json: @item
-      end
-
-
       def update
         new_email = params[:new_email]
         update_project_relationships(new_email) if new_email != @user.email
